# modules/development-host/src/main/resources/site/addzero/studio/development/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Console Client API",
    "version": "1.0.0",
    "description": "Generated from shared transport contracts."
  },
  "paths": {
...
    "/console/api/spreadsheet-templates": {
      "post": {
        "operationId": "importSpreadsheetTemplate",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "templateKey",
                  "name",
                  "file"
                ],
                "properties": {
                  "templateKey": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SpreadsheetTemplateView"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listSpreadsheetTemplates",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SpreadsheetTemplateListPage"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/console/api/spreadsheet-templates/{templateKey}": {
      "put": {
        "operationId": "updateSpreadsheetTemplate",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "templateKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpreadsheetTemplateUpdateCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SpreadsheetTemplateView"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSpreadsheetTemplate",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "templateKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getSpreadsheetTemplate",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "templateKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SpreadsheetTemplateView"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/console/api/spreadsheet-templates/{templateKey}/fill": {
      "post": {
        "operationId": "fillSpreadsheetTemplate",
        "tags": [
          "Spreadsheet Templates"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "templateKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpreadsheetTemplateFillCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Generated spreadsheet",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpreadsheetTemplateListItemView": {
        "type": "object",
        "title": "SpreadsheetTemplateListItemView",
        "required": [
          "templateKey",
          "revision",
          "name",
          "fileName",
          "macroEnabled"
        ],
        "properties": {
          "templateKey": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "macroEnabled": {
            "type": "boolean"
          }
        }
      },
      "SpreadsheetTemplateListPage": {
        "type": [
          "object",
          "null"
        ],
        "title": "SpreadsheetTemplateListPage",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetTemplateListItemView"
            }
          },
          "totalRowCount": {
            "type": "integer"
          },
          "totalPageCount": {
            "type": "integer"
          }
        }
      },
      "SpreadsheetSourceSpec": {
        "type": "object",
        "title": "SpreadsheetSourceSpec",
        "required": [
          "fileName",
          "mediaType",
          "sha256"
        ],
        "properties": {
          "fileName": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "macroEnabled": {
            "type": "boolean"
          },
          "containsExternalLinks": {
            "type": "boolean"
          }
        }
      },
      "SpreadsheetStyleSpec": {
        "type": "object",
        "title": "SpreadsheetStyleSpec",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "integer"
          },
          "fontFamily": {
            "type": [
              "string",
              "null"
            ]
          },
          "fontSizePt": {
            "type": [
              "number",
              "null"
            ]
          },
          "bold": {
            "type": "boolean"
          },
          "italic": {
            "type": "boolean"
          },
          "textColor": {
            "type": [
              "string",
              "null"
            ]
          },
          "fillColor": {
            "type": [
              "string",
              "null"
            ]
          },
          "horizontalAlignment": {
            "type": [
              "string",
              "null"
            ]
          },
          "verticalAlignment": {
            "type": [
              "string",
              "null"
            ]
          },
          "wrapText": {
            "type": "boolean"
          },
          "borderTop": {
            "type": [
              "string",
              "null"
            ]
          },
          "borderRight": {
            "type": [
              "string",
              "null"
            ]
          },
          "borderBottom": {
            "type": [
              "string",
              "null"
            ]
          },
          "borderLeft": {
            "type": [
              "string",
              "null"
            ]
          },
          "numberFormat": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SpreadsheetCellSpec": {
        "type": "object",
        "title": "SpreadsheetCellSpec",
        "required": [
          "row",
          "column",
          "type",
          "styleKey"
        ],
        "properties": {
          "row": {
            "type": "integer"
          },
          "column": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "EMPTY",
              "TEXT",
              "NUMBER",
              "BOOLEAN",
              "DATE",
              "ERROR",
              "FORMULA"
            ]
          },
          "displayValue": {
            "type": "string"
          },
          "formula": {
            "type": [
              "string",
              "null"
            ]
          },
          "styleKey": {
            "type": "integer"
          }
        }
      },
      "SpreadsheetRangeSpec": {
        "type": "object",
        "title": "SpreadsheetRangeSpec",
        "required": [
          "fromRow",
          "fromColumn"
        ],
        "properties": {
          "fromRow": {
            "type": "integer"
          },
          "fromColumn": {
            "type": "integer"
          },
          "toRow": {
            "type": "integer"
          },
          "toColumn": {
            "type": "integer"
          }
        }
      },
      "SpreadsheetImageSpec": {
        "type": "object",
        "title": "SpreadsheetImageSpec",
        "required": [
          "key",
          "mediaType",
          "dataBase64",
          "range"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "dataBase64": {
            "type": "string"
          },
          "range": {
            "$ref": "#/components/schemas/SpreadsheetRangeSpec"
          }
        }
      },
      "SpreadsheetPrintSpec": {
        "type": "object",
        "title": "SpreadsheetPrintSpec",
        "properties": {
          "area": {
            "$ref": "#/components/schemas/SpreadsheetRangeSpec"
          },
          "landscape": {
            "type": "boolean"
          },
          "paperSize": {
            "type": "integer"
          },
          "fitWidth": {
            "type": "integer"
          },
          "fitHeight": {
            "type": "integer"
          },
          "marginLeft": {
            "type": "number"
          },
          "marginRight": {
            "type": "number"
          },
          "marginTop": {
            "type": "number"
          },
          "marginBottom": {
            "type": "number"
          }
        }
      },
      "SpreadsheetSheetSpec": {
        "type": "object",
        "title": "SpreadsheetSheetSpec",
        "required": [
          "key",
          "name",
          "rowCount",
          "columnCount"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rowCount": {
            "type": "integer"
          },
          "columnCount": {
            "type": "integer"
          },
          "columnWidthsPx": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "rowHeightsPx": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "hiddenColumns": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetCellSpec"
            }
          },
          "mergedRanges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetRangeSpec"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetImageSpec"
            }
          },
          "print": {
            "$ref": "#/components/schemas/SpreadsheetPrintSpec"
          }
        }
      },
      "SpreadsheetVariableSpec": {
        "type": "object",
        "title": "SpreadsheetVariableSpec",
        "required": [
          "key",
          "label"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "NUMBER",
              "BOOLEAN",
              "DATE",
              "DATETIME",
              "IMAGE"
            ]
          },
          "required": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SpreadsheetBindingSpec": {
        "type": "object",
        "title": "SpreadsheetBindingSpec",
        "required": [
          "key",
          "variableKey",
          "sheetKey",
          "target",
          "range"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "variableKey": {
            "type": "string"
          },
          "sheetKey": {
            "type": "string"
          },
          "target": {
            "type": "string",
            "enum": [
              "CELL",
              "MERGED_RANGE",
              "IMAGE_ANCHOR"
            ]
          },
          "range": {
            "$ref": "#/components/schemas/SpreadsheetRangeSpec"
          }
        }
      },
      "SpreadsheetLedgerFieldSpec": {
        "type": "object",
        "title": "SpreadsheetLedgerFieldSpec",
        "required": [
          "key",
          "label",
          "column"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "NUMBER",
              "BOOLEAN",
              "DATE",
              "DATETIME",
              "IMAGE"
            ]
          },
          "column": {
            "type": "integer"
          },
          "required": {
            "type": "boolean"
          }
        }
      },
      "SpreadsheetLedgerSpec": {
        "type": "object",
        "title": "SpreadsheetLedgerSpec",
        "required": [
          "key",
          "name",
          "sheetKey",
          "firstRow"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sheetKey": {
            "type": "string"
          },
          "firstRow": {
            "type": "integer"
          },
          "maxRows": {
            "type": "integer"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetLedgerFieldSpec"
            }
          }
        }
      },
      "SpreadsheetCellEditSpec": {
        "type": "object",
        "title": "SpreadsheetCellEditSpec",
        "required": [
          "sheetKey",
          "row",
          "column",
          "value"
        ],
        "properties": {
          "sheetKey": {
            "type": "string"
          },
          "row": {
            "type": "integer"
          },
          "column": {
            "type": "integer"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "SpreadsheetTemplateDocument": {
        "type": "object",
        "title": "SpreadsheetTemplateDocument",
        "required": [
          "name",
          "source",
          "sheets"
        ],
        "properties": {
          "version": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "$ref": "#/components/schemas/SpreadsheetSourceSpec"
          },
          "styles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetStyleSpec"
            }
          },
          "sheets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetSheetSpec"
            }
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetVariableSpec"
            }
          },
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetBindingSpec"
            }
          },
          "ledgers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetLedgerSpec"
            }
          },
          "edits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetCellEditSpec"
            }
          }
        }
      },
      "SpreadsheetTemplateView": {
        "type": [
          "object",
          "null"
        ],
        "title": "SpreadsheetTemplateView",
        "required": [
          "templateKey",
          "revision",
          "document"
        ],
        "properties": {
          "templateKey": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "document": {
            "$ref": "#/components/schemas/SpreadsheetTemplateDocument"
          }
        }
      },
      "SpreadsheetTemplateDraftSpec": {
        "type": "object",
        "title": "SpreadsheetTemplateDraftSpec",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetVariableSpec"
            }
          },
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetBindingSpec"
            }
          },
          "ledgers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetLedgerSpec"
            }
          },
          "edits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadsheetCellEditSpec"
            }
          }
        }
      },
      "SpreadsheetTemplateUpdateCommand": {
        "type": "object",
        "title": "SpreadsheetTemplateUpdateCommand",
        "required": [
          "expectedRevision",
          "draft"
        ],
        "properties": {
          "expectedRevision": {
            "type": "integer"
          },
          "draft": {
            "$ref": "#/components/schemas/SpreadsheetTemplateDraftSpec"
          }
        }
      },
      "SpreadsheetTemplateFillCommand": {
        "type": "object",
        "title": "SpreadsheetTemplateFillCommand",
        "required": [
          "expectedRevision"
        ],
        "properties": {
          "expectedRevision": {
            "type": "integer"
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "ledgers": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}